+++ src/lib.rs
@@ -1,0 +1,10 @@
+use std::collections::HashMap;
+
+pub mod line_member;
+pub use self::line_member::LineMember;
+
+pub mod picker;
+pub use self::picker::Picker;
+
+pub type SimulationTime = f64;
+pub type ItemPicks<'a> = HashMap<&'a str, i32>;
